Refactor plugin structure and improve functionality
